backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
openpyxl==3.1.5
qrcode[pil]==7.4.2
Pillow==10.4.0
